Split C++ into multiple files
Split the relatively large iotea.cpp/hpp files in to smaller parts and
do the same for the relevant tests. This change only affects the file
structure.

Also remove the examples "function_provider" and "event_consumer". These
examples where developed as a means to test the SDK and experiment with
the platform during development. The other examples cover the
functionality demonstrated by the removed examples.

Signed-off-by: Alexander Perlman <fixed-term.Alexander.Perlman@se.bosch.com>